Use a struct representation for `mod` items
